% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dup_id_check}
\alias{dup_id_check}
\title{Check That Names are Associated with One ID}
\usage{
dup_id_check(data_check)
}
\arguments{
\item{data_check}{}
}
\value{
A table of duplicates
}
\description{
Check That Names are Associated with One ID
}
\examples{
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "electric_class.xlsx", sep = "/"))
dup_id_check(file_check)
}
