% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dup_name_check}
\alias{dup_name_check}
\title{Check that IDs are Associated with One Name}
\usage{
dup_name_check(data_check)
}
\arguments{
\item{data_check}{}
}
\value{
A table of duplicates
}
\description{
Check that IDs are Associated with One Name
}
\examples{
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "electric_class.xlsx", sep = "/"))
dup_name_check(file_check)
}
