% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_customer_sales_check.R
\name{month_customer_sales_check}
\alias{month_customer_sales_check}
\title{Electric Month Customer Sales Check}
\usage{
month_customer_sales_check(file_check, dest_dir)
}
\arguments{
\item{file_check}{A monthly sales table that needs to be checked}

\item{dest_dir}{The directory that you want the markdown file written to}
}
\value{
An HTML markdown document
}
\description{
Electric Month Customer Sales Check
}
\examples{
\dontrun{
dest_dir <- getwd()
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "electric_class_month.xlsx", sep = "/"))
class_check(file_check, dest_dir)
}
}
