% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Set Credentials}
\usage{
set_credentials()
}
\value{
Credentials for logging into the data warehouse
}
\description{
Set Credentials
}
\examples{
\dontrun{
set_credentials()
}
}
