% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_heating_check.R
\name{space_heating_check}
\alias{space_heating_check}
\title{Electric Space Heating Check}
\usage{
space_heating_check(file_check, dest_dir)
}
\arguments{
\item{file_check}{A county electricity table that needs to be checked}

\item{dest_dir}{The directory that you want the markdown file written to}
}
\value{
An HTML markdown document
}
\description{
Electric Space Heating Check
}
\examples{
\dontrun{
dest_dir <- getwd()
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "electric_class.xlsx", sep = "/"))
class_check(file_check, dest_dir)
}
}
