% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{valid_id_check}
\alias{valid_id_check}
\title{Utility ID check}
\usage{
valid_id_check(data_check, utility_table)
}
\arguments{
\item{data_check}{A dataset to check}

\item{utility_table}{A list of valid utility IDs}
}
\value{
A table of non-valid utility IDs
}
\description{
Utility ID check
}
\examples{
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "electric_class.xlsx", sep = "/"))
id_table <- tibble::tribble(~utility_id, "some_id")
valid_name_check(file_check, id_table)
}
