% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{valid_name_check}
\alias{valid_name_check}
\title{Utility Name Check}
\usage{
valid_name_check(data_check, utility_table)
}
\arguments{
\item{data_check}{A dataset to check}

\item{utility_table}{A list of valid utility names}
}
\value{
A table of non-valid utility names
}
\description{
Utility Name Check
}
\examples{
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "electric_class.xlsx", sep = "/"))
name_table <- tibble::tribble(~utility_name, "some_name")
valid_name_check(file_check, name_table)
}
