% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{county_check}
\alias{county_check}
\title{County Check}
\usage{
county_check(file_check, dest_dir)
}
\arguments{
\item{file_check}{A county electricity table that needs to be checked}

\item{dest_dir}{The directory that you want the markdown file written to}
}
\value{
An HTML markdown document
}
\description{
County Check
}
