% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_check.R
\name{county_check}
\alias{county_check}
\title{County Check}
\usage{
county_check(file_check, dest_dir)
}
\arguments{
\item{file_check}{A county electricity table that needs to be checked}

\item{dest_dir}{The directory that you want the markdown file written to}
}
\value{
An HTML markdown document
}
\description{
County Check
}
\examples{
\dontrun{
dest_dir <- getwd()
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "county_data.xlsx", sep = "/"))
county_check(file_check, dest_dir)
}
}
