% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{county_check}
\alias{county_check}
\title{County Check}
\usage{
county_check(file_path)
}
\arguments{
\item{file_path}{The path to the county electricity file that needs to be checked}
}
\value{
An HTML markdown document
}
\description{
County Check
}
