% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dup_check}
\alias{dup_check}
\title{Title}
\usage{
dup_check(data_check)
}
\arguments{
\item{data_check}{Data set that is being checked}
}
\value{
A table of duplicates
}
\description{
Title
}
\examples{
dir <- system.file("extdata", package = "elecchecks")
file_check <- readxl::read_xlsx(paste(dir, "county_data.xlsx", sep = "/"))
dup_check(file_check)
}
